src/core/requirements.txt
discord.py==1.3.3
ptvsd==4.3.2
SQLAlchemy==1.3.17
PyMySQL==0.9.3